Factor out test library
