+++ test/Jamfile.v2
@@ -2,7 +2,7 @@
 
 # Copyright (c) 2003, 2006 Beman Dawes
 # Copyright (c) 2012 Artyom Beilis (Tonkikh)
-# Copyright (c) 2020 Alexander Grund
+# Copyright (c) 2020-2021 Alexander Grund
 # 
 # Distributed under the Boost Software License, Version 1.0.
 # (See accompanying file LICENSE or www.boost.org/LICENSE_1_0.txt)
@@ -35,22 +35,25 @@ project : requirements
   ;
 
 lib shell32 ;
+lib file_test_helpers : file_test_helpers.cpp : <link>static -<library>/boost/nowide//boost_nowide ;
 
 run test_codecvt.cpp ;
 run test_convert.cpp ;
 run test_env.cpp ;
 run test_env.cpp : : : <define>BOOST_NOWIDE_TEST_INCLUDE_WINDOWS=1 : test_env_win ;
 run test_fs.cpp : : : <library>/boost/filesystem//boost_filesystem/<warnings-as-errors>off ;
-run test_filebuf.cpp  ;
-run test_filebuf.cpp : : : <define>BOOST_NOWIDE_USE_FILEBUF_REPLACEMENT=1 <target-os>windows:<build>no : test_internal_filebuf ;
-run test_ifstream.cpp ;
-run test_ofstream.cpp ;
-run test_fstream.cpp ;
-run test_fstream.cpp : : : <define>BOOST_NOWIDE_USE_FILEBUF_REPLACEMENT=1 <target-os>windows:<build>no : test_internal_fstream ;
-run test_fstream_cxx11.cpp  ;
-run test_fstream_cxx11.cpp : : : <define>BOOST_NOWIDE_USE_FILEBUF_REPLACEMENT=1 <target-os>windows:<build>no : test_internal_fstream_cxx11 ;
-run test_fstream_special.cpp  ;
-run test_fstream_special.cpp : : : <define>BOOST_NOWIDE_USE_FILEBUF_REPLACEMENT=1 <target-os>windows:<build>no : test_internal_fstream_special ;
+run test_filebuf.cpp file_test_helpers ;
+run test_filebuf.cpp file_test_helpers : : : <define>BOOST_NOWIDE_USE_FILEBUF_REPLACEMENT=1 <target-os>windows:<build>no : test_filebuf_internal ;
+run test_ifstream.cpp file_test_helpers ;
+run test_ifstream.cpp file_test_helpers : : : <define>BOOST_NOWIDE_USE_FILEBUF_REPLACEMENT=1 <target-os>windows:<build>no : test_ifstream_internal ;
+run test_ofstream.cpp file_test_helpers ;
+run test_ofstream.cpp file_test_helpers : : : <define>BOOST_NOWIDE_USE_FILEBUF_REPLACEMENT=1 <target-os>windows:<build>no : test_ofstream_internal ;
+run test_fstream.cpp file_test_helpers ;
+run test_fstream.cpp file_test_helpers : : : <define>BOOST_NOWIDE_USE_FILEBUF_REPLACEMENT=1 <target-os>windows:<build>no : test_fstream_internal ;
+run test_fstream_cxx11.cpp file_test_helpers ;
+run test_fstream_cxx11.cpp file_test_helpers : : : <define>BOOST_NOWIDE_USE_FILEBUF_REPLACEMENT=1 <target-os>windows:<build>no : test_fstream_cxx11_internal ;
+run test_fstream_special.cpp file_test_helpers ;
+run test_fstream_special.cpp file_test_helpers : : : <define>BOOST_NOWIDE_USE_FILEBUF_REPLACEMENT=1 <target-os>windows:<build>no : test_fstream_special_internal ;
 run test_iostream.cpp ;
 run test_stackstring.cpp ;
 run test_stat.cpp ;
